refactor(cors): convert int64 to int

+++ internal/domain/types.go
@@ -9,6 +9,6 @@ type CORSConfig struct {
 	AllowedMethods   []string
 	AllowedHeaders   []string
 	ExposedHeaders   []string
-	MaxAge           int64
+	MaxAge           int
 	AllowCredentials bool
 }
